requirements.txt.txt
Flask==2.1.0
torch==1.10.1
matplotlib==3.5.1
numpy==1.21.5
Pillow==8.4.0
torchvision==0.11.2
